vista para comprar, vender y realizar alertas de BTC
